fix: reattach a stolen doc comment and bind the carry-forward cutoff
Both from review.

The new constant and method were inserted between clean_up_prices and
its doc comment, so "Remove all prices older than the latest price date"
ended up documenting CARRY_FORWARD_MAX_AGE_DAYS and clean_up_prices was
left undocumented. Rust doc comments attach to the next item, so this
was silently wrong in both directions.

max_age_days is a value, so it is bound rather than interpolated, per
the convention documented on ConnectionPool::count. The table names stay
interpolated because identifiers cannot be bound - that is the existing
delete_prices_before_latest pattern, not an exception to it. Typed i32
because Postgres defines date - int4, not date - int8; verified the
bound query runs against real Postgres rather than only compiling.

+++ src/price/service.rs
@@ -259,7 +259,6 @@ impl PriceService {
         Ok(())
     }
 
-    /// Remove all prices older than the latest price date, in one statement (§5).
     /// How far back to look for a price to carry forward.
     ///
     /// Chosen to match the retention tiers - daily for 7 days, weekly to 28,
@@ -267,7 +266,7 @@ impl PriceService {
     /// roughly a month a frozen number is more misleading than an absent one:
     /// a card that stopped trading should eventually read as having no price
     /// rather than holding whatever it last sold for.
-    pub const CARRY_FORWARD_MAX_AGE_DAYS: i64 = 30;
+    pub const CARRY_FORWARD_MAX_AGE_DAYS: i32 = 30;
 
     /// Fill in cards today's build did not mention, from the newest price we
     /// have for them. See [`PriceRepository::carry_forward_missing_prices`] for
@@ -280,6 +279,7 @@ impl PriceService {
         Ok(carried)
     }
 
+    /// Remove all prices older than the latest price date, in one statement (§5).
     pub async fn clean_up_prices(&self) -> Result<()> {
         let deleted = self.repository.delete_prices_before_latest().await?;
         if deleted == 0 {
